tests/dev_requirements.txt
flake8 ~= 3.7
pytest ~= 4.6
pytest-cov ~= 2.8
pytest-mock ~= 2.0
lazy-import ~= 0.2
pytz == 2019.3
pydataclasses ~= 1.0
